fix(docs): swizzle NotFound/Content to fix DocRoot generic-404 flash
DocRoot renders @theme/NotFound/Content directly when no doc route
matches an unknown URL — not @theme/NotFound (the SPA catch-all).
Because <Route path="/"> matches every path before the * catch-all in
the Switch, the generic theme-classic NotFoundContent ("We could not
find what you were looking for.") was always shown after SPA boot,
overwriting the correct static 404.html.

Fix: swizzle @theme/NotFound/Content with our branded 404 content
(no Layout wrapper, as DocsRoot already provides one). Restructure
@theme/NotFound into a directory so both index.tsx and Content/ can
coexist; index.tsx now delegates to NotFoundContent rather than
duplicating markup.
